+++ src/hooks/card/useCardAlias.ts
@@ -1,4 +1,5 @@
-import { useCardDispatch, useCardState } from "./useCard";
+import { useCardDispatch, useCardState } from "../../contexts/CardContext";
+
 
 function useCardAlias() {
   const dispatch = useCardDispatch();
