+++ src/hooks/card/useCardOwnerName.ts
@@ -1,4 +1,4 @@
-import { useCardDispatch, useCardState } from "./useCard";
+import { useCardDispatch, useCardState } from "../../contexts/CardContext";
 
 function useCardOwnerName() {
   const dispatch = useCardDispatch();
